Renamed postgres folder files
