+++ postgresdb/nonrelationaldb/seedNonRelationalDB.js
@@ -6,11 +6,11 @@ const password = process.env.POSTGRES_PW || 'marzagat';
 const port = parseInt(process.env.POSTGRES_PORT, 10) || 5432;
 const batchSize = process.env.BATCH_SIZE || 10000;
 const numRecords = process.env.NUM_RECORDS || 10000000;
-const database = process.env.DATABASE || 'business_info_postgres_relational';
+const database = process.env.DATABASE || 'business_info_postgres_non_relational';
 
 const pgp = require('pg-promise')({});
 
-const generateData = require('./generateDataPSDB');
+const makeFakeData = require('./generateNonRelationalData');
 
 const cn = `postgres://${username}:${password}@${host}:${port}/${database}`;
 
@@ -24,11 +24,11 @@ const connectDB = async () => {
     });
 };
 
-const insertRestaurantBatch = async (restaurants) => {
-  const restaurantColumns = ['place_id', 'formatted_address', 'international_phone_number', 'url', 'lat', 'lng', 'weekday_text'];
+const insertRestaurantBatch = async (fakeRestaurants) => {
+  const restaurantColumns = ['place_id', 'formatted_address', 'international_phone_number', 'url', 'lat', 'lng', 'weekday_text', 'monday_open', 'monday_close', 'tuesday_open', 'tuesday_close', 'wednesday_open', 'wednesday_close', 'thursday_open', 'thursday_close', 'friday_open', 'friday_close', 'saturday_open', 'saturday_close', 'sunday_open', 'sunday_close'];
   const csRestaurants = new pgp.helpers.ColumnSet(restaurantColumns, { table: 'restaurants' });
 
-  const insertionQueryRestaurants = pgp.helpers.insert(restaurants, csRestaurants);
+  const insertionQueryRestaurants = pgp.helpers.insert(fakeRestaurants, csRestaurants);
 
   await db.none(insertionQueryRestaurants)
     .catch((err) => {
@@ -36,31 +36,14 @@ const insertRestaurantBatch = async (restaurants) => {
     });
 };
 
-const insertHoursBatch = async (fakeHours) => {
-  const hoursColumns = ['restaurant_id', 'weekday', 'open_time', 'close_time'];
-  const csHours = new pgp.helpers.ColumnSet(hoursColumns, { table: 'hours' });
-  const insertionQueryHours = pgp.helpers.insert(fakeHours, csHours);
-
-  await db.none(insertionQueryHours)
-    .catch((err) => {
-      console.error(err);
-    });
-};
-
 const insertBatch = async (startId) => {
   const restaurantsBatch = [];
-  const hoursBatch = [];
   for (let i = 0; i < batchSize; i += 1) {
-    const { restaurantRow, hoursRows } = generateData(startId + i);
-    restaurantsBatch.push(restaurantRow);
-    
-    for (let j = 0; j < hoursRows.length; j += 1) {
-      hoursBatch.push(hoursRows[j]);
-    }
+    const { fakeRestaurantRow } = makeFakeData(startId + i);
+    restaurantsBatch.push(fakeRestaurantRow);
   }
 
   await insertRestaurantBatch(restaurantsBatch);
-  await insertHoursBatch(hoursBatch);
 };
 
 const seedDb = async (numRecords) => {
@@ -83,23 +66,15 @@ const indexRestaurants = async () => {
     .catch(error => console.error(error));
 };
 
-const addForeignKeys = async () => {
-  const foreignKeyQuery = 'ALTER TABLE hours ADD CONSTRAINT fk FOREIGN KEY (restaurant_id) REFERENCES restaurants(place_id)';
-  await db.none(foreignKeyQuery)
-    .then(() => console.log('added foreign keys'))
-    .catch(error => console.error(error));
-};
-
 const startSeed = Date.now();
 
 seedDb(numRecords)
   .then(() => console.log('seeded!!'))
   .then(() => indexRestaurants())
-  .then(() => addForeignKeys())
   .then(() => {
     const endSeed = Date.now();
     console.log('Seed time: ', endSeed - startSeed, 'ms');
   })
   .catch(err => console.error(err));
-
+  
 module.exports = db;
